requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pyyaml
requests
pandas
sentence-transformers
esco-skill-extractor
wbgapi
python-multipart
google-generativeai
python-dotenv
